Make functions cache values
I am not fond of the new `store()` functionality as it requires a
synchronous approach to where specific cached information is set and
used.

And I understand how it can help with the image-processing
functionality, however I think this mostly is a display of bad practices
and actually requires a redesign.

This PR returns back to the original "if cached, return cache, otherwise
perform lookup" for a simple case, but I'd like to see a different
approach to the other uses.